requirements.txt
asn1crypto==0.24.0
atomicwrites==1.1.5
attrs==18.1.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
coverage==4.5.1
cryptography==1.9
idna==2.7
jwt==0.5.3
more-itertools==4.2.0
Playoff==0.7.1
pluggy==0.6.0
py==1.5.3
pycparser==2.18
PyJWT==1.6.4
pytest==3.6.2
python-dotenv==0.8.2
requests==2.19.1
selenium==3.12.0
six==1.11.0
urllib3==1.23
